requirements.txt
streamlit
sqlalchemy
python-dotenv
langchain
openai